Ajoute des methodes programEventDeplacement,programEventRemplir,programEventDeversement dans robot.java - Amelioration de liste d'evenement dans simulateur.java pour faire paralellism - J'ai eu mal a factoriser le code - make exeSimulationSujet pour voir la resultat

+++ src/robot/entities/Incendie.java
@@ -50,7 +50,7 @@ public class Incendie extends Entity {
 	return nbLitresEauPourExtinction;
     }
 
-    public void nbLitresEauArrive(int vol){
+    public void nbLitresEauArrive(double vol){
 	this.nbLitresEauPourExtinction -= vol;
     }
